Conflicts:
	src/engine.c
	src/runner.h
	src/runner_doiact.h
	src/scheduler.c

+++ src/drift.h
@@ -39,7 +39,7 @@
  * @param ti_current Integer end of time-step
  */
 __attribute__((always_inline)) INLINE static void drift_gpart(
-    struct gpart *restrict gp, float dt, double timeBase, integertime_t ti_old,
+    struct gpart *restrict gp, double dt, double timeBase, integertime_t ti_old,
     integertime_t ti_current) {
 
 #ifdef SWIFT_DEBUG_CHECKS
@@ -75,7 +75,7 @@ __attribute__((always_inline)) INLINE static void drift_gpart(
  * @param ti_current Integer end of time-step
  */
 __attribute__((always_inline)) INLINE static void drift_part(
-    struct part *restrict p, struct xpart *restrict xp, float dt,
+    struct part *restrict p, struct xpart *restrict xp, double dt,
     double timeBase, integertime_t ti_old, integertime_t ti_current) {
 
 #ifdef SWIFT_DEBUG_CHECKS
@@ -119,7 +119,7 @@ __attribute__((always_inline)) INLINE static void drift_part(
  * @param ti_current Integer end of time-step
  */
 __attribute__((always_inline)) INLINE static void drift_spart(
-    struct spart *restrict sp, float dt, double timeBase, integertime_t ti_old,
+    struct spart *restrict sp, double dt, double timeBase, integertime_t ti_old,
     integertime_t ti_current) {
 
 #ifdef SWIFT_DEBUG_CHECKS
